[BEAM-77] Move hadoop contrib into hdfs IO
