+++ sdks/java/io/hdfs/pom.xml
@@ -23,36 +23,43 @@
 
   <parent>
     <groupId>org.apache.beam</groupId>
-    <artifactId>sdks-parent</artifactId>
+    <artifactId>io-parent</artifactId>
     <version>0.1.0-incubating-SNAPSHOT</version>
     <relativePath>../pom.xml</relativePath>
   </parent>
 
-  <artifactId>java-sdk-parent</artifactId>
+  <artifactId>hdfs</artifactId>
+  <name>Apache Beam :: SDKs :: Java :: IO :: HDFS</name>
+  <description>Library to read and write Hadoop/HDFS file formats from Beam.</description>
+  <packaging>jar</packaging>
 
-  <packaging>pom</packaging>
+  <dependencies>
+    <dependency>
+      <groupId>org.apache.beam</groupId>
+      <artifactId>java-sdk-all</artifactId>
+      <version>${project.version}</version>
+    </dependency>
 
-  <name>Apache Beam :: SDKs :: Java</name>
+    <dependency>
+      <groupId>org.apache.hadoop</groupId>
+      <artifactId>hadoop-client</artifactId>
+      <version>2.7.0</version>
+      <scope>provided</scope>
+    </dependency>
 
-  <modules>
-    <module>core</module>
-    <module>io</module>
-    <!-- sdks/java/maven-archtypes has several dependencies on the
-         DataflowPipelineRunner. Until these are refactored out or
-         a released artifact exists, we need to modify the build order.
-    <module>maven-archetypes</module> -->
-  </modules>
+    <!-- test dependencies -->
+    <dependency>
+      <groupId>org.hamcrest</groupId>
+      <artifactId>hamcrest-all</artifactId>
+      <version>1.3</version>
+      <scope>test</scope>
+    </dependency>
 
-  <profiles>
-    <profile>
-      <id>java8-tests</id>
-      <activation>
-        <jdk>[1.8,)</jdk>
-      </activation>
-      <modules>
-        <module>java8tests</module>
-      </modules>
-    </profile>
-  </profiles>
-
+    <dependency>
+      <groupId>junit</groupId>
+      <artifactId>junit</artifactId>
+      <version>4.11</version>
+      <scope>test</scope>
+    </dependency>
+  </dependencies>
 </project>
