+++ sdks/java/io/hdfs/src/main/java/org/apache/beam/sdk/io/hdfs/HDFSFileSource.java
@@ -15,7 +15,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.beam.contrib.hadoop;
+package org.apache.beam.sdk.io.hdfs;
 
 import org.apache.beam.sdk.coders.Coder;
 import org.apache.beam.sdk.coders.KvCoder;
@@ -53,32 +53,32 @@ import java.util.NoSuchElementException;
 import javax.annotation.Nullable;
 
 /**
- * A {@code BoundedSource} for reading files resident in a Hadoop filesystem using a
+ * A {@code BoundedSource} for reading files resident in a Hadoop filesystem (HDFS) using a
  * Hadoop file-based input format.
  *
  * <p>To read a {@link org.apache.beam.sdk.values.PCollection} of
  * {@link org.apache.beam.sdk.values.KV} key-value pairs from one or more
- * Hadoop files, use {@link HadoopFileSource#from} to specify the path(s) of the files to
+ * Hadoop files, use {@link HDFSFileSource#from} to specify the path(s) of the files to
  * read, the Hadoop {@link org.apache.hadoop.mapreduce.lib.input.FileInputFormat}, the
  * key class and the value class.
  *
- * <p>A {@code HadoopFileSource} can be read from using the
+ * <p>A {@code HDFSFileSource} can be read from using the
  * {@link org.apache.beam.sdk.io.Read} transform. For example:
  *
  * <pre>
  * {@code
- * HadoopFileSource<K, V> source = HadoopFileSource.from(path, MyInputFormat.class,
+ * HDFSFileSource<K, V> source = HDFSFileSource.from(path, MyInputFormat.class,
  *   MyKey.class, MyValue.class);
  * PCollection<KV<MyKey, MyValue>> records = Read.from(mySource);
  * }
  * </pre>
  *
- * <p>The {@link HadoopFileSource#readFrom} method is a convenience method
+ * <p>The {@link HDFSFileSource#readFrom} method is a convenience method
  * that returns a read transform. For example:
  *
  * <pre>
  * {@code
- * PCollection<KV<MyKey, MyValue>> records = HadoopFileSource.readFrom(path,
+ * PCollection<KV<MyKey, MyValue>> records = HDFSFileSource.readFrom(path,
  *   MyInputFormat.class, MyKey.class, MyValue.class);
  * }
  * </pre>
@@ -91,7 +91,7 @@ import javax.annotation.Nullable;
  * @param <K> The type of keys to be read from the source.
  * @param <V> The type of values to be read from the source.
  */
-public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
+public class HDFSFileSource<K, V> extends BoundedSource<KV<K, V>> {
   private static final long serialVersionUID = 0L;
 
   private final String filepattern;
@@ -101,7 +101,7 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
   private final SerializableSplit serializableSplit;
 
   /**
-   * Creates a {@code Read} transform that will read from an {@code HadoopFileSource}
+   * Creates a {@code Read} transform that will read from an {@code HDFSFileSource}
    * with the given file name or pattern ("glob") using the given Hadoop
    * {@link org.apache.hadoop.mapreduce.lib.input.FileInputFormat},
    * with key-value types specified by the given key class and value class.
@@ -112,34 +112,34 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
   }
 
   /**
-   * Creates a {@code HadoopFileSource} that reads from the given file name or pattern ("glob")
+   * Creates a {@code HDFSFileSource} that reads from the given file name or pattern ("glob")
    * using the given Hadoop {@link org.apache.hadoop.mapreduce.lib.input.FileInputFormat},
    * with key-value types specified by the given key class and value class.
    */
-  public static <K, V, T extends FileInputFormat<K, V>> HadoopFileSource<K, V> from(
+  public static <K, V, T extends FileInputFormat<K, V>> HDFSFileSource<K, V> from(
       String filepattern, Class<T> formatClass, Class<K> keyClass, Class<V> valueClass) {
     @SuppressWarnings("unchecked")
-    HadoopFileSource<K, V> source = (HadoopFileSource<K, V>)
-        new HadoopFileSource(filepattern, formatClass, keyClass, valueClass);
+    HDFSFileSource<K, V> source = (HDFSFileSource<K, V>)
+        new HDFSFileSource(filepattern, formatClass, keyClass, valueClass);
     return source;
   }
 
   /**
-   * Create a {@code HadoopFileSource} based on a file or a file pattern specification.
+   * Create a {@code HDFSFileSource} based on a file or a file pattern specification.
    */
-  private HadoopFileSource(String filepattern,
-      Class<? extends FileInputFormat<?, ?>> formatClass, Class<K> keyClass,
-      Class<V> valueClass) {
+  private HDFSFileSource(String filepattern,
+                         Class<? extends FileInputFormat<?, ?>> formatClass, Class<K> keyClass,
+                         Class<V> valueClass) {
     this(filepattern, formatClass, keyClass, valueClass, null);
   }
 
   /**
-   * Create a {@code HadoopFileSource} based on a single Hadoop input split, which won't be
+   * Create a {@code HDFSFileSource} based on a single Hadoop input split, which won't be
    * split up further.
    */
-  private HadoopFileSource(String filepattern,
-      Class<? extends FileInputFormat<?, ?>> formatClass, Class<K> keyClass,
-      Class<V> valueClass, SerializableSplit serializableSplit) {
+  private HDFSFileSource(String filepattern,
+                         Class<? extends FileInputFormat<?, ?>> formatClass, Class<K> keyClass,
+                         Class<V> valueClass, SerializableSplit serializableSplit) {
     this.filepattern = filepattern;
     this.formatClass = formatClass;
     this.keyClass = keyClass;
@@ -166,13 +166,13 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
   @Override
   public void validate() {
     Preconditions.checkNotNull(filepattern,
-        "need to set the filepattern of a HadoopFileSource");
+        "need to set the filepattern of a HDFSFileSource");
     Preconditions.checkNotNull(formatClass,
-        "need to set the format class of a HadoopFileSource");
+        "need to set the format class of a HDFSFileSource");
     Preconditions.checkNotNull(keyClass,
-        "need to set the key class of a HadoopFileSource");
+        "need to set the key class of a HDFSFileSource");
     Preconditions.checkNotNull(valueClass,
-        "need to set the value class of a HadoopFileSource");
+        "need to set the value class of a HDFSFileSource");
   }
 
   @Override
@@ -183,7 +183,7 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
           new Function<InputSplit, BoundedSource<KV<K, V>>>() {
         @Nullable @Override
         public BoundedSource<KV<K, V>> apply(@Nullable InputSplit inputSplit) {
-          return new HadoopFileSource<K, V>(filepattern, formatClass, keyClass,
+          return new HDFSFileSource<K, V>(filepattern, formatClass, keyClass,
               valueClass, new SerializableSplit(inputSplit));
         }
       });
@@ -212,9 +212,9 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
     this.validate();
 
     if (serializableSplit == null) {
-      return new HadoopFileReader<>(this, filepattern, formatClass);
+      return new HDFSFileReader<>(this, filepattern, formatClass);
     } else {
-      return new HadoopFileReader<>(this, filepattern, formatClass,
+      return new HDFSFileReader<>(this, filepattern, formatClass,
           serializableSplit.getSplit());
     }
   }
@@ -269,7 +269,7 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
     return false;
   }
 
-  static class HadoopFileReader<K, V> extends BoundedSource.BoundedReader<KV<K, V>> {
+  static class HDFSFileReader<K, V> extends BoundedSource.BoundedReader<KV<K, V>> {
 
     private final BoundedSource<KV<K, V>> source;
     private final String filepattern;
@@ -284,18 +284,18 @@ public class HadoopFileSource<K, V> extends BoundedSource<KV<K, V>> {
     private KV<K, V> currentPair;
 
     /**
-     * Create a {@code HadoopFileReader} based on a file or a file pattern specification.
+     * Create a {@code HDFSFileReader} based on a file or a file pattern specification.
      */
-    public HadoopFileReader(BoundedSource<KV<K, V>> source, String filepattern,
-        Class<? extends FileInputFormat<?, ?>> formatClass) {
+    public HDFSFileReader(BoundedSource<KV<K, V>> source, String filepattern,
+                          Class<? extends FileInputFormat<?, ?>> formatClass) {
       this(source, filepattern, formatClass, null);
     }
 
     /**
-     * Create a {@code HadoopFileReader} based on a single Hadoop input split.
+     * Create a {@code HDFSFileReader} based on a single Hadoop input split.
      */
-    public HadoopFileReader(BoundedSource<KV<K, V>> source, String filepattern,
-        Class<? extends FileInputFormat<?, ?>> formatClass, InputSplit split) {
+    public HDFSFileReader(BoundedSource<KV<K, V>> source, String filepattern,
+                          Class<? extends FileInputFormat<?, ?>> formatClass, InputSplit split) {
       this.source = source;
       this.filepattern = filepattern;
       this.formatClass = formatClass;
